refactor: move nfdc to package
