+++ tools/nfdc/nfdc_dataset.go
@@ -1,4 +1,4 @@
-package tools
+package nfdc
 
 import (
 	"fmt"
