+++ tools/nfdc/nfdc_route.go
@@ -1,4 +1,4 @@
-package tools
+package nfdc
 
 import (
 	"fmt"
