+++ tools/nfdc/nfdc_status.go
@@ -1,4 +1,4 @@
-package tools
+package nfdc
 
 import (
 	"fmt"
